Create gameWithIntegers.java
Vanya and Vova are playing a game. Players are given an integer 𝑛
. On their turn, the player can add 1
 to the current integer or subtract 1
. The players take turns; Vanya starts. If after Vanya's move the integer is divisible by 3
, then he wins. If 10
 moves have passed and Vanya has not won, then Vova wins.

Write a program that, based on the integer 𝑛
, determines who will win if both players play optimally.

Input
The first line contains the integer 𝑡
 (1≤𝑡≤100
) — the number of test cases.

The single line of each test case contains the integer 𝑛
 (1≤𝑛≤1000
).

Output
For each test case, print "First" without quotes if Vanya wins, and "Second" without quotes if Vova wins.