Create howMuchDoesDaytonaCost.java
We define an integer to be the most common on a subsegment, if its number of occurrences on that subsegment is larger than the number of occurrences of any other integer in that subsegment. A subsegment of an array is a consecutive segment of elements in the array 𝑎
.

Given an array 𝑎
 of size 𝑛
, and an integer 𝑘
, determine if there exists a non-empty subsegment of 𝑎
 where 𝑘
 is the most common element.

Input
Each test consists of multiple test cases. The first line contains a single integer 𝑡
 (1≤𝑡≤1000
) — the number of test cases. The description of test cases follows.

The first line of each test case contains two integers 𝑛
 and 𝑘
 (1≤𝑛≤100
, 1≤𝑘≤100
) — the number of elements in array and the element which must be the most common.

The second line of each test case contains 𝑛
 integers 𝑎1
, 𝑎2
, 𝑎3
, …
, 𝑎𝑛
 (1≤𝑎𝑖≤100
) — elements of the array.

Output
For each test case output "YES" if there exists a subsegment in which 𝑘
 is the most common element, and "NO" otherwise.

You can output the answer in any case (for example, the strings "yEs", "yes", "Yes", and "YES" will be recognized as a positive answer).