Create goalsOfVictory.java
There are 𝑛
 teams in a football tournament. Each pair of teams match up once. After every match, Pak Chanek receives two integers as the result of the match, the number of goals the two teams score during the match. The efficiency of a team is equal to the total number of goals the team scores in each of its matches minus the total number of goals scored by the opponent in each of its matches.

After the tournament ends, Pak Dengklek counts the efficiency of every team. Turns out that he forgot about the efficiency of one of the teams. Given the efficiency of 𝑛−1
 teams 𝑎1,𝑎2,𝑎3,…,𝑎𝑛−1
. What is the efficiency of the missing team? It can be shown that the efficiency of the missing team can be uniquely determined.

Input
Each test contains multiple test cases. The first line contains an integer 𝑡
 (1≤𝑡≤500
) — the number of test cases. The following lines contain the description of each test case.

The first line contains a single integer 𝑛
 (2≤𝑛≤100
) — the number of teams.

The second line contains 𝑛−1
 integers 𝑎1,𝑎2,𝑎3,…,𝑎𝑛−1
 (−100≤𝑎𝑖≤100
) — the efficiency of 𝑛−1
 teams.

Output
For each test case, output a line containing an integer representing the efficiency of the missing team.